+++ packages/transformer/src/IModelTransformer.ts
@@ -23,7 +23,14 @@ import {
   YieldManager,
 } from "@itwin/core-bentley";
 import * as ECSchemaMetaData from "@itwin/ecschema-metadata";
-import { Point3d, Transform } from "@itwin/core-geometry";
+import {
+  Angle,
+  AxisIndex,
+  Matrix3d,
+  Point3d,
+  Transform,
+  Vector3d,
+} from "@itwin/core-geometry";
 import {
   BriefcaseManager,
   ChangedECInstance,
@@ -79,7 +86,9 @@ import {
   EntityReference,
   ExternalSourceAspectProps,
   FontProps,
+  GeometricElement3dProps,
   GeometricElementProps,
+  Helmert2DWithZOffset,
   IModel,
   IModelError,
   ModelProps,
@@ -238,6 +247,14 @@ export interface IModelTransformOptions {
    * @default undefined
    */
   argsForProcessChanges?: ProcessChangesOptions;
+
+  /**
+   * A flag that determines if spatial elements from the source db should be transformed if:
+   * source and target iModel GCS/CRS data is the same, but they have differing additional transforms
+   * source and target iModel ECEF locations differ
+   * @default false
+   */
+  tryAlignGeolocation?: boolean;
 }
 
 /**
@@ -378,6 +395,15 @@ export class IModelTransformer extends IModelExportHandler {
   public readonly targetDb: IModelDb;
   /** The IModelTransformContext for this IModelTransformer. */
   public readonly context: IModelCloneContext;
+  /** The transform to be applied to the placement of spatial elements
+   * This transform should be applied when:
+   * - source and target db have different ECEF locations
+   * - source and target db have matching GCS/CRS data, but differing `geographicCoordinateSystem.additionalTransform.helmert2DWithZOffset`
+   * @note for ECEF transforms, this can only be used when source and target are linearly located imodels
+   * @note for non linearly located imodels, this transform will be a linear transform derived from Helmert Transforms from the src and target iModels.
+   * @beta
+   */
+  private _linearSpatialTransform?: Transform;
   private _syncType?: SyncType;
 
   /** The Id of the Element in the **target** iModel that represents the **source** repository as a whole and scopes its [ExternalSourceAspect]($backend) instances. */
@@ -585,6 +611,7 @@ export class IModelTransformer extends IModelExportHandler {
         options?.branchRelationshipDataBehavior ?? "reject",
       skipPropagateChangesToRootElements:
         options?.skipPropagateChangesToRootElements ?? true,
+      tryAlignGeolocation: options?.tryAlignGeolocation ?? false,
     };
     // check if authorization client is defined
     if (IModelHost.authorizationClient === undefined) {
@@ -652,6 +679,29 @@ export class IModelTransformer extends IModelExportHandler {
       (this.targetDb as any).codeValueBehavior = "exact";
     }
     /* eslint-enable @itwin/no-internal */
+    if (this._options.tryAlignGeolocation) {
+      if (
+        this.sourceDb.geographicCoordinateSystem ||
+        this.targetDb.geographicCoordinateSystem
+      ) {
+        Logger.logTrace(
+          loggerCategory,
+          "Aligning Additional transforms between imodels due to imodels containing GeographicCoordinateSystem data"
+        );
+        this._linearSpatialTransform =
+          this.calculateTransformFromHelmertTransforms();
+      } else if (this.sourceDb.ecefLocation && this.targetDb.ecefLocation) {
+        Logger.logTrace(
+          loggerCategory,
+          "Aligning ECEF Location's between imodels due to imodels not containing GeographicCoordinateSystem data"
+        );
+        this._linearSpatialTransform = this.calculateEcefTransform();
+      } else
+        Logger.logTrace(
+          loggerCategory,
+          "No Geolcation data to align, both GCS and ECEF are undefined"
+        );
+    }
   }
 
   /** validates that the importer set on the transformer has the same values for its shared options as the transformer.
@@ -1553,7 +1603,162 @@ export class IModelTransformer extends IModelExportHandler {
         targetElementProps.jsonProperties.Subject.Job = undefined;
       }
     }
+
+    if (
+      this._linearSpatialTransform !== undefined &&
+      sourceElement instanceof GeometricElement3d
+    ) {
+      // can check the sourceElement since this IModelTransformer does not remap classes
+      const placement = Placement3d.fromJSON(
+        (targetElementProps as GeometricElement3dProps).placement
+      );
+
+      if (placement.isValid) {
+        placement.multiplyTransform(this._linearSpatialTransform);
+        (targetElementProps as GeometricElement3dProps).placement = placement;
+      }
+    }
     return targetElementProps;
+  }
+
+  /**
+   * Calculate the transform between two ECEF locations
+   * @param srcDb
+   * @param targetDb
+   * @returns Transform that converts relative coordinates in the source iModel to relative coordinates in the target iModel.
+   * @note This can only be used if both source and target iModels are linearly located
+   */
+  public calculateEcefTransform(): Transform | undefined {
+    const srcEcefLoc = this.sourceDb.ecefLocation;
+    const targetEcefLoc = this.targetDb.ecefLocation;
+
+    if (srcEcefLoc === undefined || targetEcefLoc === undefined) {
+      throw new IModelError(
+        IModelStatus.NoGeoLocation,
+        "Both source and target ECEF locations must be defined to calculate the transform."
+      );
+    }
+    if (srcEcefLoc.getTransform().isAlmostEqual(targetEcefLoc.getTransform())) {
+      Logger.logTrace(
+        loggerCategory,
+        "ECEF data is already aligned. No spatial transforms needed."
+      );
+      return undefined;
+    }
+
+    const srcSpatialToECEF = srcEcefLoc.getTransform(); // converts relative to ECEF in relation to source
+    const targetECEFToSpatial = targetEcefLoc.getTransform().inverse(); // converts ECEF to relative in relation to target
+    if (!targetECEFToSpatial) {
+      throw new IModelError(
+        IModelStatus.NoGeoLocation,
+        "Failed to invert target ECEF transform."
+      );
+    }
+    const ecefTransform =
+      targetECEFToSpatial.multiplyTransformTransform(srcSpatialToECEF); // chain both transforms
+
+    return ecefTransform;
+  }
+
+  public static convertHelmertToTransform(
+    helmert: Helmert2DWithZOffset | undefined
+  ): Transform {
+    if (!helmert) {
+      return Transform.createIdentity();
+    }
+
+    const rotationXY = Matrix3d.createRotationAroundAxisIndex(
+      AxisIndex.Z,
+      Angle.createDegrees(helmert?.rotDeg)
+    );
+    rotationXY.scaleColumnsInPlace(helmert.scale, helmert.scale, 1.0);
+    const translation = Vector3d.create(
+      helmert.translationX,
+      helmert.translationY,
+      helmert.translationZ
+    );
+    const helmertTransform = Transform.createRefs(translation, rotationXY);
+
+    return helmertTransform;
+  }
+
+  public calculateTransformFromHelmertTransforms(): Transform | undefined {
+    if (
+      this.sourceDb.geographicCoordinateSystem?.horizontalCRS === undefined ||
+      this.sourceDb.geographicCoordinateSystem?.verticalCRS === undefined
+    ) {
+      throw new IModelError(
+        IModelStatus.BadRequest,
+        "Source iModel does not have a geographic coordinate system defined."
+      );
+    }
+    if (
+      this.targetDb.geographicCoordinateSystem?.horizontalCRS === undefined ||
+      this.targetDb.geographicCoordinateSystem.verticalCRS === undefined
+    ) {
+      throw new IModelError(
+        IModelStatus.BadRequest,
+        "Target iModel does not have a geographic coordinate system defined."
+      );
+    }
+    if (
+      !this.sourceDb.geographicCoordinateSystem.horizontalCRS.equals(
+        this.targetDb.geographicCoordinateSystem.horizontalCRS
+      ) ||
+      !this.sourceDb.geographicCoordinateSystem.verticalCRS.equals(
+        this.targetDb.geographicCoordinateSystem.verticalCRS
+      )
+    ) {
+      throw new IModelError(
+        IModelStatus.MismatchGcs,
+        "Source and target geographic coordinate systems must match to calculate the spatial transform."
+      );
+    }
+    if (
+      this.sourceDb.geographicCoordinateSystem.additionalTransform ===
+      this.targetDb.geographicCoordinateSystem.additionalTransform
+    ) {
+      Logger.logTrace(
+        loggerCategory,
+        "Geolocation data is already aligned. No spatial transforms needed."
+      );
+      return undefined;
+    }
+
+    const srcScale =
+      this.sourceDb.geographicCoordinateSystem.additionalTransform
+        ?.helmert2DWithZOffset?.scale ?? 1;
+    const targetScale =
+      this.targetDb.geographicCoordinateSystem.additionalTransform
+        ?.helmert2DWithZOffset?.scale ?? 1;
+
+    if (srcScale !== targetScale) {
+      throw new IModelError(
+        IModelStatus.MismatchGcs,
+        "Spatial transform is non rigid. Source and target Helmert transforms must have the same scale to calculate a rigid spatial transform."
+      );
+    }
+
+    const srcTransform = IModelTransformer.convertHelmertToTransform(
+      this.sourceDb.geographicCoordinateSystem.additionalTransform
+        ?.helmert2DWithZOffset
+    ); // moves elements to where src helmert transform would move them at render time
+    const targetTransformInv = IModelTransformer.convertHelmertToTransform(
+      this.targetDb.geographicCoordinateSystem.additionalTransform
+        ?.helmert2DWithZOffset
+    ).inverse(); // negates target helmert transform that is applied at render time
+
+    if (!targetTransformInv) {
+      throw new IModelError(
+        IModelStatus.NoGeoLocation,
+        "Failed to invert target Helmert transform."
+      );
+    }
+
+    const combinedTransform =
+      targetTransformInv.multiplyTransformTransform(srcTransform);
+
+    return combinedTransform;
   }
 
   // if undefined, it can be initialized by calling [[this.processChangesets]]